requirements/base.txt
# Core
Django==3.2

# Filter
django-filter==2.4.0

# Django template
django-mathfilters==1.0.0

# Configuration
django-split-settings==1.2.0

# Media
django-image-cropping==1.5.0
django-widget-tweaks==1.4.8
easy-thumbnails==2.7.1
Markdown==3.3.3
Pillow==9.3.0

# REST
djangorestframework==3.12.2

# DB
redis==3.5.3

# Env
python-decouple==3.3

# Utils
six==1.15.0
